ig-traces: Fix TraceIcon pod object reference
The mentioned component had not been updated when we started using pods
as Pod class instances, but the code to exec into the pod was expecting
it to be a Pod class instance. So the icon functionality for viewing
the traces was completely broken.

This patch fixes that by simply using the pod object as a Pod instance.